Add test for overview with data (supposed to fail at the moment)
