+++ tests/gui/test_04_preprocessing.py
@@ -1,8 +1,8 @@
 from alphastats import DataSet
-from alphastats.load_data import load_data
 from streamlit.testing.v1 import AppTest
 from pathlib import Path
 from unittest.mock import MagicMock, patch
+from .test_helper import create_dataset_alphapept
 import pandas as pd
 from io import BytesIO
 
@@ -10,25 +10,6 @@ from io import BytesIO
 APP_FOLDER = Path(__file__).parent / Path("../../alphastats/gui/")
 APP_FILE = f"{APP_FOLDER}/pages/03_Preprocessing.py"
 TEST_INPUT_FILES = f"{APP_FOLDER}/../../testfiles"
-
-
-def print_session_state(apptest: AppTest):
-    for k, v in apptest.session_state.filtered_state.items():
-        print(
-            f"{k}:    {str(type(v))}   {str(v)[:20] if type(v) not in [int, list, str] else v}"
-        )
-
-
-def create_dataset_alphapept():
-    loader = load_data(
-        file=TEST_INPUT_FILES + "/alphapept/results_proteins.csv", type="alphapept"
-    )
-    metadata_path = TEST_INPUT_FILES + "/alphapept/metadata.csv"
-    return DataSet(
-        loader=loader,
-        metadata_path=metadata_path,
-        sample_column="sample",
-    )
 
 
 def test_page_04_loads_without_input():
